Fixed bug where coverage was always 100%
The inclusion of `require 'simplecov'` lines in the gem caused simplecov to
shift focus and concider the gem to be the entity under test instead of the
including project. Since the gem ha no test and runs no lines the coverage
always ends up being 100%.
Removing all refrences to RSpec and SimpleCov from the gem and relying on the
including project to require them (since using this gem would be pointless if it
doesn't) solves the problem and makes the gem work as expected.